Move compare page components into separate directories, based on their tab
