+++ site/frontend/src/pages/compare/compile/filters.vue
