+++ site/frontend/src/pages/compare/compile/quick-links.vue
